pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dim-sequencer"
version = "0.5.0"
description = "D.I.M — Dawless Is More: a non-linear performance sequencer for human musicians"
authors = [
    { name = "Olivier Bareau", email = "olivier.bareau@gmail.com" },
]
license = { text = "MIT" }
readme = "README.md"
requires-python = ">=3.11"
keywords = ["music", "performance", "sequencer", "live", "ableton-link", "midi"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Multimedia :: Sound/Audio",
    "Topic :: Multimedia :: Sound/Audio :: MIDI",
]

dependencies = [
    "flask>=3.0,<4",
    "flask-socketio>=5.0",
    "python-osc>=1.8",
    "zeroconf>=0.80",
    "textual>=0.50",
    "python-socketio[client]>=5.0",
]

[project.optional-dependencies]
midi = ["python-rtmidi>=1.5"]
dev  = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "black>=24.0",
    "ruff>=0.4",
]

[project.urls]
Homepage   = "https://github.com/obareau/dim"
Repository = "https://github.com/obareau/dim"
Issues     = "https://github.com/obareau/dim/issues"

[project.scripts]
dim = "dim_pkg.__main__:main"

[tool.setuptools.packages.find]
where   = ["."]
include = ["dim_pkg*", "core*", "network*", "adapters*"]

[tool.setuptools.package-data]
"adapters.web" = ["templates/**/*", "static/**/*"]
"formats"      = ["*.json"]

# ── pytest ────────────────────────────────────────────────────────────────────
[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths    = ["tests"]

# ── ruff ─────────────────────────────────────────────────────────────────────
[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "W", "I"]
ignore = ["E501"]

# ── black ────────────────────────────────────────────────────────────────────
[tool.black]
line-length = 100
target-version = ["py311"]
